# src/swagger/swagger.yaml
openapi: 3.0.3
info:
  title: TopAgents API
  description: |
    ## TopAgents.ca — Real Estate Agent Review & Referral Platform

    ### Authentication
    Protected endpoints require a `Bearer` JWT token in the `Authorization` header.
    Get tokens via `POST /api/v1/auth/login`. Refresh via `POST /api/v1/auth/refresh`.

    ### Rate Limiting
    - General: 100 req / 15 min
    - Auth endpoints: 10 req / 15 min
    - Search: 60 req / 1 min

    ### Pagination
    All list endpoints accept `?page=1&limit=20`.
    Response includes `meta: { total, page, limit, totalPages, hasNext, hasPrev }`.

    ### Response Envelope
    ```json
    { "success": true, "data": { ... } }
    ```
    ```json
    { "success": false, "error": { "code": 422, "message": "Validation failed", "details": [] } }
    ```
  version: 1.0.0
  contact:
    name: TopAgents Dev Team
    email: dev@topagents.ca
servers:
  - url: http://localhost:3000/api/v1
    description: Local development
  - url: https://api.topagents.ca/api/v1
    description: Production
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
        hasNext:
          type: boolean
        hasPrev:
          type: boolean
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: integer
            message:
              type: string
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string
...
    Agent:
      type: object
      properties:
        id:
          type: integer
        license_number:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        email:
          type: string
          format: email
          nullable: true
        phone:
          type: string
          nullable: true
        brokerage:
          type: string
          nullable: true
        bio:
          type: string
          nullable: true
        photo_url:
          type: string
          nullable: true
        website:
          type: string
          nullable: true
        languages:
          type: array
          items:
            type: string
        specialties:
          type: array
          items:
            type: string
        areas_served:
          type: array
          items:
            type: string
        claimed_at:
          type: string
          format: date-time
          nullable: true
        verified_at:
          type: string
          format: date-time
          nullable: true
        current_tier:
          type: string
          enum: [BRONZE, SILVER, GOLD, PLATINUM, DIAMOND, LUMINARY]
        total_points:
          type: integer
        year_points:
          type: integer
        avg_rating:
          type: number
          format: float
        total_reviews:
          type: integer
        connected_platforms:
          type: integer
        completed_referrals:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    AgentCreate:
      type: object
      required: [license_number, first_name, last_name]
      properties:
        license_number:
          type: string
          maxLength: 50
        first_name:
          type: string
          maxLength: 100
        last_name:
          type: string
          maxLength: 100
        email:
          type: string
          format: email
        phone:
          type: string
        brokerage:
          type: string
        bio:
          type: string
          maxLength: 2000
        website:
          type: string
          format: uri
        languages:
          type: array
          items:
            type: string
        specialties:
          type: array
          items:
            type: string
        areas_served:
          type: array
          items:
            type: string
...
  responses:
    Unauthorized:
      description: Missing or invalid JWT token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Insufficient role permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
paths:
...
  /agents:
    get:
      tags: [Agents]
      summary: Search and filter agents
      parameters:
        - in: query
          name: q
          schema:
            type: string
          description: Search by name or brokerage
        - in: query
          name: city
          schema:
            type: string
        - in: query
          name: province
          schema:
            type: string
        - in: query
          name: min_rating
          schema:
            type: number
            minimum: 1
            maximum: 5
        - in: query
          name: tier
          schema:
            type: string
            enum: [BRONZE, SILVER, GOLD, PLATINUM, DIAMOND, LUMINARY]
        - in: query
          name: specialties
          schema:
            type: string
          description: Comma-separated list
        - in: query
          name: languages
          schema:
            type: string
          description: Comma-separated list
        - in: query
          name: sort
          schema:
            type: string
            enum: [rating, reviews, recent, points]
            default: rating
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        '200':
          description: Paginated list of agents
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Agent'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'

    post:
      tags: [Agents]
      summary: Create agent profile (admin or bulk seed)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentCreate'
      responses:
        '201':
          description: Agent created
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: License number already exists
...
  /admin/dashboard:
    get:
      tags: [Admin]
      summary: Platform KPI overview
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Counts of agents, reviews, users, leads, referrals, flagged reviews
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      agents:
                        type: object
                        properties:
                          total:
                            type: integer
                          claimed:
                            type: integer
                      reviews:
                        type: object
                        properties:
                          total:
                            type: integer
                          verified:
                            type: integer
                      users:
                        type: object
                        properties:
                          total:
                            type: integer
                          agents:
                            type: integer
                      leads:
                        type: object
                        properties:
                          total:
                            type: integer
                          new_leads:
                            type: integer
                      referrals:
                        type: object
                        properties:
                          total:
                            type: integer
                          completed:
                            type: integer
                      flagged_reviews:
                        type: object
                        properties:
                          total:
                            type: integer
  /admin/agents:
    get:
      tags: [Admin]
      summary: List all agents including unclaimed profiles
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [claimed, unclaimed, all]
            default: all
        - in: query
          name: q
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Paginated full agent list
  /admin/reviews/flagged:
    get:
      tags: [Admin]
      summary: Get flagged reviews pending moderation
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Paginated flagged review reports
  /admin/agents/{agentId}/suspend:
    patch:
      tags: [Admin]
      summary: Suspend or reactivate an agent profile
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: agentId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [suspended]
              properties:
                suspended:
                  type: boolean
                reason:
                  type: string
      responses:
        '200':
          description: Agent suspended or reactivated
  /admin/tiers/recalculate:
    post:
      tags: [Admin]
      summary: Manually trigger nightly tier recalculation for all agents
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Tier recalculation complete